backend/requirements.txt
arxiv == 1.4.8
anthropic==0.3.6
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
distro==1.8.0
exceptiongroup==1.1.2
fastapi==0.100.1
firebase-admin==6.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pydantic==1.10.12
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
tokenizers==0.13.3
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1